Refactor authentication-related validation to minimize Teamy API calls
