+++ backend/src/middlewares/fishtank.js
@@ -1,26 +1,26 @@
 const { checkSchema } = require('express-validator/check');
-const Auth = require('./validation/customRules/checkers/auth');
+const User = require('./validation/customRules/checkers/user');
 const ValidationSchema = require('./validation/schemas/fishtank.js');
 const checkCustomRules = require('./validation/customRules/checkers/fishtank.js');
 const bail = require('./validation/errorHandler.js');
 
 module.exports = {
   create: [
-    Auth.isAuthenticated,
+    User.isAuthenticated,
     bail(403),
     checkSchema(ValidationSchema.create),
     bail(422),
-    Auth.isValidShoal,
+    User.isValidShoal,
     bail(422),
   ],
   edit: [
-    Auth.isAuthenticated,
+    User.isAuthenticated,
     bail(403),
     checkSchema(ValidationSchema.edit),
     bail(422),
     checkCustomRules.edit,
     bail(422),
-    Auth.isOwner,
+    User.isOwner,
     bail(403),
   ],
 };
